+++ Changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+### Added
+ - Description support for fields.
+ - WYSIWYG editor for the text area component to use Quill.js
+
 ## 2.11.6
 ### Fixed
  - Another instance where infinite JSON structures could occur with eachComponent.